NUKE EVERYTHINGGGGG. jk there was weird duplicate project within itself?????

+++ src/VitaWave.Common/APIToWebserver/PersonPoint.cs
@@ -12,8 +12,8 @@ namespace VitaWave.Common.APIToWebserver
         public const string UNKNOWN_STATUS = "Unknown";
         public const uint UNKNOWN_TID = 255;
 
-        public required double X { get; set; }
-        public required double Y { get; set; }
+        public required float X { get; set; }
+        public required float Y { get; set; }
         public uint TID { get; set; } = UNKNOWN_TID; // Unknown
         public string Status { get; set; } = UNKNOWN_STATUS;
     }
